+++ macros/src/main/scala/fr/gstraymond/parser/PriceModels.scala
@@ -10,7 +10,7 @@ object PriceModels {
 
   case class CardPricePartial(paper: Option[Price], online: Option[Price])
 
-  case class Price(normal: Option[Float], foil: Option[Float])
+  case class Price(normal: Option[Double], foil: Option[Double])
 
   implicit val cardPriceFormat: JsonValueCodec[Map[String, CardPricePartial]] =
     JsonCodecMaker.make(CodecMakerConfig.withTransientEmpty(false).withMapMaxInsertNumber(1_000_000))
